+++ Realm/Realm.h
@@ -32,10 +32,10 @@
 #import <Realm/RLMSyncConfiguration.h>
 #import <Realm/RLMSyncCredentials.h>
 #import <Realm/RLMSyncManager.h>
+#import <Realm/RLMSyncPermission.h>
 #import <Realm/RLMSyncPermissionOffer.h>
 #import <Realm/RLMSyncPermissionOfferResponse.h>
 #import <Realm/RLMSyncPermissionResults.h>
-#import <Realm/RLMSyncPermissionValue.h>
 #import <Realm/RLMSyncSession.h>
 #import <Realm/RLMSyncUser.h>
 #import <Realm/RLMSyncUtil.h>