+++ CHANGELOG.md
@@ -52,7 +52,8 @@ x.x.x Release notes (yyyy-MM-dd)
 * Computed properties on Realm object types are detected and no
   longer added to the automatically generated schema.
 * Remove deprecated `{RLM}SyncPermission` and `{RLM}SyncPermissionChange`
-  classes. Use `{RLM}SyncPermissionValue` instead.
+  classes.
+* `{RLM}SyncPermissionValue` has been renamed to just `{RLM}SyncPermission`.
 * Remove deprecated `-[RLMSyncUser permissionRealmWithError:]` and
   `SyncUser.permissionRealm()` APIs. Use the new permissions system.
 * Remove deprecated error `RLMSyncErrorBadResponse`. Use